Finalized SpeedBoost and ScoreMult Pop Up
I finished the animations for the VFX, imported a new font which I applied to our score and coin text and resized the score tracker and its BG. Also added options menu and updated map select.

Animations are in the Hierarchy, hidden with a Canvas Group on the Animation object ready for Mauro to implement.